refactor: update join or create room modal
